optimize local variable lookup with u16, down to 4080ms in fibo34

+++ src/calcit/symbol.rs
@@ -86,7 +86,7 @@ pub struct CalcitImport {
   /// from npm package, use `default` and asterisk in js
   pub def: Arc<str>,
   pub info: Arc<ImportInfo>,
-  pub coord: Option<(usize, usize)>,
+  pub coord: Option<(u16, u16)>,
 }
 
 /// compare at namespace level, ignore at_def
